fix: realtime RPS/TPS float64, RPM from backend directly
- Backend: RPS/TPS int64 -> float64 to avoid integer truncation at low traffic
- Frontend: RPM uses statsRealtime.rpm instead of rps*60
- Frontend: TPS displays 1 decimal place
- Update tests for float64

+++ internal/model/realtime_stats.go
@@ -2,9 +2,9 @@ package model
 
 // StatsRealtime 实时监控指标（基于滑动窗口，不持久化）
 type StatsRealtime struct {
-	WindowSizeSec int   `json:"window_size_sec"`
-	RPS           int64 `json:"rps"`
-	RPM           int64 `json:"rpm"`
-	TPS           int64 `json:"tps"`
-	TPM           int64 `json:"tpm"`
+	WindowSizeSec int     `json:"window_size_sec"`
+	RPS           float64 `json:"rps"`
+	RPM           int64   `json:"rpm"`
+	TPS           float64 `json:"tps"`
+	TPM           int64   `json:"tpm"`
 }
